refactor: move api to items module
