+++ src/app/modules/items/api/services/api-items/api-items.service.spec.ts
@@ -2,9 +2,9 @@ import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController, TestRequest } from '@angular/common/http/testing';
 import { ApiItemsService } from './api-items.service';
 import { RouterTestingModule } from '@angular/router/testing';
-import { environment } from '../../../../../environments/environment';
-import { ItemContract } from './contracts/item.contract';
-import { itemMockContract } from './mocks/item-mock.contract';
+import { environment } from '../../../../../../environments/environment';
+import { ItemContract } from '../../contracts/item.contract';
+import { itemMockContract } from '../../mocks/item-mock.contract';
 import { IApiResponse } from '../../interfaces/response.interface';
 import { API_ITEMS_URL } from '../../constants/api.constants';
 import { HttpHeaders, HttpResponse } from '@angular/common/http';
