+++ src/app/modules/items/services/items.facade.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { select, Store } from '@ngrx/store';
-import { IQueryParams } from '../../api/interfaces/pagination.interface';
+import { IQueryParams } from '@api';
 import { getItems, resetStateItems } from '../state/items.actions';
 import { ItemsState } from '../state/items.reducer';
 import { ItemModel } from '../models/item.model';
