+++ src/app/modules/items/services/items/items.service.ts
@@ -1,12 +1,9 @@
 import { Injectable } from '@angular/core';
-import { ApiItemsService } from '@api';
-import { IQueryParams } from '../../../api/interfaces/pagination.interface';
 import { ItemModel } from '../../models/item.model';
-import { ItemContract } from '../../../api/services/api-items/contracts/item.contract';
 import { map } from 'rxjs/operators';
-import { IApiResponse } from '@api';
 import { Observable } from 'rxjs';
 import { ItemTranslator } from '../../translate/item.translator';
+import { IApiResponse, ApiItemsService, IQueryParams, ItemContract } from '@api';
 
 @Injectable({
   providedIn: 'root'
