+++ src/app/modules/items/state/items.effects.ts
@@ -3,11 +3,10 @@ import { of } from 'rxjs';
 import { map, catchError, exhaustMap } from 'rxjs/operators';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { HttpErrorResponse } from '@angular/common/http';
-import { IApiResponse } from '@api';
 import { getItems, getItemsSuccess, getItemsFailure } from './items.actions';
 import { ItemsService } from '../services/items/items.service';
 import { ItemModel } from '../models/item.model';
-import { IQueryParams } from '../../api/interfaces/pagination.interface';
+import { IQueryParams, IApiResponse } from '@api';
 
 
 @Injectable()
